Initial Upload
This is the inital upload to github. Includes code that captures chat alongside viewers. This should be branched off as the focus of the project has been changed.

Current Bug: login does not work properly. Meetme has changed the way they handle login requests. Need to figure out what change that is.